Limit dashboard interactions to components in the same group
